+++ glfw/x11_window.c
@@ -3384,6 +3384,17 @@ _glfwPlatformSetWindowBlur(_GLFWwindow *window, int blur_radius) {
 }
 
 
+bool
+_glfwPlatformGrabKeyboard(bool grab) {
+    int result;
+    if (grab) {
+        result = XGrabKeyboard(_glfw.x11.display, _glfw.x11.root, True, GrabModeAsync, GrabModeAsync, CurrentTime);
+    } else {
+        result = XUngrabKeyboard(_glfw.x11.display, CurrentTime);
+    }
+    return result == GrabSuccess;
+}
+
 //////////////////////////////////////////////////////////////////////////
 //////                        GLFW native API                       //////
 //////////////////////////////////////////////////////////////////////////