+++ kittens/icat/native.go
@@ -4,13 +4,13 @@ package icat
 
 import (
 	"fmt"
+	"github.com/kovidgoyal/kitty/tools/tty"
+	"github.com/kovidgoyal/kitty/tools/tui/graphics"
+	"github.com/kovidgoyal/kitty/tools/utils"
+	"github.com/kovidgoyal/kitty/tools/utils/images"
+	"github.com/kovidgoyal/kitty/tools/utils/shm"
 	"image"
 	"image/gif"
-	"kitty/tools/tty"
-	"kitty/tools/tui/graphics"
-	"kitty/tools/utils"
-	"kitty/tools/utils/images"
-	"kitty/tools/utils/shm"
 
 	"github.com/edwvee/exiffix"
 	"github.com/kovidgoyal/imaging"