+++ kittens/icat/transmit.go
@@ -8,20 +8,20 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"github.com/kovidgoyal/kitty"
 	"io"
-	"kitty"
 	"math"
 	not_rand "math/rand/v2"
 	"os"
 	"path/filepath"
 	"strings"
 
-	"kitty/tools/tui"
-	"kitty/tools/tui/graphics"
-	"kitty/tools/tui/loop"
-	"kitty/tools/utils"
-	"kitty/tools/utils/images"
-	"kitty/tools/utils/shm"
+	"github.com/kovidgoyal/kitty/tools/tui"
+	"github.com/kovidgoyal/kitty/tools/tui/graphics"
+	"github.com/kovidgoyal/kitty/tools/tui/loop"
+	"github.com/kovidgoyal/kitty/tools/utils"
+	"github.com/kovidgoyal/kitty/tools/utils/images"
+	"github.com/kovidgoyal/kitty/tools/utils/shm"
 )
 
 var _ = fmt.Print