+++ kittens/themes/ui.go
@@ -4,20 +4,20 @@ package themes
 
 import (
 	"fmt"
+	"github.com/kovidgoyal/kitty"
 	"io"
-	"kitty"
 	"maps"
 	"regexp"
 	"slices"
 	"strings"
 	"time"
 
-	"kitty/tools/config"
-	"kitty/tools/themes"
-	"kitty/tools/tui/loop"
-	"kitty/tools/tui/readline"
-	"kitty/tools/utils"
-	"kitty/tools/wcswidth"
+	"github.com/kovidgoyal/kitty/tools/config"
+	"github.com/kovidgoyal/kitty/tools/themes"
+	"github.com/kovidgoyal/kitty/tools/tui/loop"
+	"github.com/kovidgoyal/kitty/tools/tui/readline"
+	"github.com/kovidgoyal/kitty/tools/utils"
+	"github.com/kovidgoyal/kitty/tools/wcswidth"
 )
 
 var _ = fmt.Print