+++ kitty/simple_cli_definitions.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # License: GPLv3 Copyright: 2025, Kovid Goyal <kovid at kovidgoyal.net>
 
-# This module must be runnable by a vanilla python interperter
+# This module must be runnable by a vanilla python interpreter
 # as it is used to generate C code when building kitty
 
 import re
@@ -323,6 +323,14 @@ def generate_c_parsers() -> Iterator[str]:
 
 
 # kitty CLI spec {{{
+grab_keyboard_docs = """\
+Grab the keyboard. This means global shortcuts defined in the OS will be passed to kitty instead. Useful if
+you want to create an OS modal window. How well this
+works depends on the OS/window manager/desktop environment. On Wayland it works only if the compositor implements
+the :link:`inhibit-keyboard-shortcuts protocol <https://wayland.app/protocols/keyboard-shortcuts-inhibit-unstable-v1>`.
+On macOS Apple doesn't allow applications to grab the keyboard without special permissions, so it doesn't work.
+"""
+
 listen_on_defn = f'''\
 --listen-on
 completion=type:special group:complete_kitty_listen_on
@@ -375,12 +383,12 @@ def kitty_options_spec() -> str:
 dest=cls
 default={appname}
 condition=not is_macos
-Set the :italic:`application id` on Wayland. On X11 set the class part of the :italic:`WM_CLASS` window property.
+On Wayland set the :italic:`application id`. On X11 set the class part of the :italic:`WM_CLASS` window property.
 
 
 --name --os-window-tag
 condition=not is_macos
-On Wayland, set the :italic:`window tag`.
+On Wayland, set the :italic:`window tag`, when specified.
 On X11, set the name part of the :italic:`WM_CLASS` property, when unset, defaults to using the
 value from :option:`{appname} --class`.
 
@@ -484,6 +492,11 @@ See also :opt:`remember_window_position` to have kitty automatically try
 to restore the previous window position.
 
 
+--grab-keyboard
+type=bool-set
+{grab_keyboard_docs}
+
+
 # Debugging options
 
 --version -v
@@ -539,8 +552,9 @@ type=bool-set
 '''
         setattr(kitty_options_spec, 'ans', OPTIONS.format(
             appname=appname, conf_name=appname, listen_on_defn=listen_on_defn,
-            config_help=CONFIG_HELP.format(appname=appname, conf_name=appname),
-        ))
+            grab_keyboard_docs=grab_keyboard_docs,
+            config_help=CONFIG_HELP.format(appname=appname, conf_name=appname
+        )))
     ans: str = getattr(kitty_options_spec, 'ans')
     return ans
 # }}}
@@ -650,7 +664,7 @@ Use the special value :code:`list` to get a list of available outputs.
 dest=cls
 default={cls}
 condition=not is_macos
-Set the :italic:`application id` on Wayland. On X11 set the class part of the :italic:`WM_CLASS` window property.
+On Wayland set the :italic:`namespace` of the layer shell surface. On X11 set the class part of the :italic:`WM_CLASS` window property.
 
 
 --name --os-window-tag
@@ -673,6 +687,11 @@ type=bool-set
 Automatically hide the panel window when it loses focus. Using this option will force :option:`--focus-policy`
 to :code:`on-demand`. Note that on Wayland, depending on the compositor, this can result in the window never
 becoming visible.
+
+
+--grab-keyboard
+type=bool-set
+{grab_keyboard_docs}
 
 
 --exclusive-zone
@@ -740,7 +759,7 @@ Path to a log file to store STDOUT/STDERR when using :option:`--detach`
 --debug-rendering
 type=bool-set
 For internal debugging use.
-'''.format(appname=appname, listen_on_defn=listen_on_defn, **d)
+'''.format(appname=appname, listen_on_defn=listen_on_defn, grab_keyboard_docs=grab_keyboard_docs, **d)
 
 
 # }}}