+++ tools/cmd/at/send_text.go
@@ -5,11 +5,11 @@ package at
 import (
 	"encoding/base64"
 	"errors"
+	"github.com/kovidgoyal/kitty/tools/tty"
+	"github.com/kovidgoyal/kitty/tools/tui/loop"
+	"github.com/kovidgoyal/kitty/tools/utils"
+	"github.com/kovidgoyal/kitty/tools/utils/shlex"
 	"io"
-	"kitty/tools/tty"
-	"kitty/tools/tui/loop"
-	"kitty/tools/utils"
-	"kitty/tools/utils/shlex"
 	"os"
 	"strings"
 )